requirements.txt
django~=4.2.3
psycopg2
djangorestframework~=3.14.0
products~=1.2.0